The 'Ternary' package allows the creation of ternary plots (a.k.a. ternary graphs / simplex plots / Gibbs triangles / de Finetti diagrams) using the familiar functions of the default 'graphics' package.

For simple use cases, generate ternary plots using the point-and-click Shiny app:

install.packages('Ternary')
Ternary::TernaryApp()

Details

For greater control over your plots, use the R command line; for usage instructions, consult the guidance article online or by typing vignette('Ternary', 'Ternary') at the R console.

See also

Author

Maintainer: Martin R. Smith martin.smith@durham.ac.uk (ORCID) [copyright holder]

Other contributors:

Lilian Sanselme [contributor]