Cartesian coordinates to ternary point

Convert cartesian (x, y) coordinates to a point in ternary space.

XYToTernary(x, y, direction = getOption("ternDirection"))

Arguments

x, y
Numeric values giving the x and y coordinates of a point or points.

direction
(optional) Integer specifying the direction that the current ternary plot should point: 1, up; 2, right; 3, down; 4, left.

Value

XYToTernary() Returns the ternary point(s) corresponding to the specified x and y coordinates, where a + b + c = 1.

See also

Examples

XYToTernary(c(0.1, 0.2), 0.5)

#>        [,1]       [,2]
#> a 0.5773503 0.57735027
#> b 0.3113249 0.41132487
#> c 0.1113249 0.01132487